refactor(utils): put functions in separate files
